+++ js/clickHandler.js
@@ -6,4 +6,5 @@ var clickCount = document.getElementById("clickCount");
 document.getElementById("clicker").addEventListener("click", function() {
     clicks += perClick;
     document.getElementById("clickCount").innerText = `Clicks: ${clicks}`;
+    playSound("click");
 });