+++ js/soundHandler.js
@@ -1,0 +1,3 @@
+function playSound(sound) {
+    var sound = new Audio(`./sounds/${sound}.mp3`).play();
+}