+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Modern Some Web Clicker [0.6.0 - 0.9.0]
+### Unreleased [10/25/2018]
+#### Additions
+- Added click and buy sounds.
+
+#### Known Bugs
+- Changing the clicker button text breaks the ripple.
+
 ### 0.9.0 [10/24/2018]
 #### Additions
 - **NEW!** Auto Clicker upgrade.